feat: Introduce flush_helpers for agent execution buffer management
- Added `flush_helpers.py` to provide shared functions for parsing traces, marking processing timeouts, and writing traces to S3 in parallel, facilitating better management of agent execution buffers.
- Refactored `flush_executions.py` and `sweep_stuck_executions.py` to utilize the new helper functions, improving code organization and reducing duplication.
- Updated `buffer.py` to expose a shared traces storage instance, enhancing accessibility across different components.
