+++ retail/agents/domains/agent_execution/usecases/sweep_stuck_executions.py
@@ -14,14 +14,22 @@ from uuid import UUID
 
 from django.conf import settings
 from django.utils import timezone
+from redis import Redis
+from redis.exceptions import RedisError
 
+from retail.agents.domains.agent_execution.flush_helpers import (
+    mark_processing_as_timed_out,
+    parse_traces,
+    write_traces_parallel,
+)
 from retail.agents.domains.agent_execution.models import (
     AgentExecution,
     AgentExecutionStatus,
 )
+from retail.agents.domains.agent_execution.services import buffer as _buffer_module
 from retail.agents.domains.agent_execution.services.buffer import (
     ExecutionBufferService,
-    _get_shared_traces_storage,
+    get_shared_traces_storage,
 )
 from retail.agents.domains.agent_execution.services.traces_storage import (
     ExecutionTracesStorageService,
@@ -40,51 +48,45 @@ class SweepStuckExecutionsUseCase:
         stuck_threshold_seconds: Optional[int] = None,
         batch_size: Optional[int] = None,
         s3_parallel_puts: Optional[int] = None,
+        redis_client: Optional[Redis] = None,
     ):
         self._traces_storage = traces_storage
-        self.stuck_threshold_seconds = (
-            stuck_threshold_seconds
-            if stuck_threshold_seconds is not None
-            else getattr(
-                settings,
-                "AGENT_EXECUTION_STUCK_THRESHOLD_SECONDS",
-                ExecutionBufferService.DEFAULT_STUCK_THRESHOLD_SECONDS,
-            )
+        self._redis_client = redis_client
+        self.stuck_threshold_seconds = self._resolve(
+            stuck_threshold_seconds,
+            "AGENT_EXECUTION_STUCK_THRESHOLD_SECONDS",
+            ExecutionBufferService.DEFAULT_STUCK_THRESHOLD_SECONDS,
         )
-        self.batch_size = (
-            batch_size
-            if batch_size is not None
-            else getattr(
-                settings,
-                "AGENT_EXECUTION_FLUSH_BATCH_SIZE",
-                ExecutionBufferService.DEFAULT_FLUSH_BATCH_SIZE,
-            )
+        self.batch_size = self._resolve(
+            batch_size,
+            "AGENT_EXECUTION_FLUSH_BATCH_SIZE",
+            ExecutionBufferService.DEFAULT_FLUSH_BATCH_SIZE,
         )
-        self.s3_parallel_puts = (
-            s3_parallel_puts
-            if s3_parallel_puts is not None
-            else getattr(
-                settings,
-                "AGENT_EXECUTION_S3_PARALLEL_PUTS",
-                ExecutionBufferService.DEFAULT_S3_PARALLEL_PUTS,
-            )
+        self.s3_parallel_puts = self._resolve(
+            s3_parallel_puts,
+            "AGENT_EXECUTION_S3_PARALLEL_PUTS",
+            ExecutionBufferService.DEFAULT_S3_PARALLEL_PUTS,
         )
+
+    @staticmethod
+    def _resolve(explicit: Optional[int], setting_name: str, default: int) -> int:
+        if explicit is not None:
+            return explicit
+        return getattr(settings, setting_name, default)
 
     @property
     def traces_storage(self) -> ExecutionTracesStorageService:
         if self._traces_storage is None:
-            self._traces_storage = _get_shared_traces_storage()
+            self._traces_storage = get_shared_traces_storage()
         return self._traces_storage
 
-    def execute(self, redis_client) -> int:
-        # Imported lazily to avoid a circular import between
-        # flush_executions and this module.
-        from retail.agents.domains.agent_execution.usecases.flush_executions import (
-            _parse_traces,
-            mark_processing_as_timed_out,
-            write_traces_parallel,
-        )
+    @property
+    def redis_client(self) -> Redis:
+        if self._redis_client is None:
+            self._redis_client = _buffer_module.get_redis_connection("default")
+        return self._redis_client
 
+    def execute(self) -> int:
         cutoff = timezone.now() - timedelta(seconds=self.stuck_threshold_seconds)
         stuck = list(
             AgentExecution.objects.filter(
@@ -95,6 +97,7 @@ class SweepStuckExecutionsUseCase:
         if not stuck:
             return 0
 
+        redis_client = self.redis_client
         s3_writes: List[Tuple[UUID, str, List[Dict[str, Any]]]] = []
         cleanup_keys: List[str] = []
         for execution_uuid, traces_s3_key in stuck:
@@ -104,7 +107,7 @@ class SweepStuckExecutionsUseCase:
                 traces_raw = redis_client.lrange(
                     ExecutionBufferService.traces_key(execution_uuid), 0, -1
                 )
-            except Exception:
+            except RedisError:
                 traces_raw = []
             if traces_raw:
                 key = traces_s3_key or self.traces_storage.get_traces_key(
@@ -114,7 +117,7 @@ class SweepStuckExecutionsUseCase:
                     (
                         execution_uuid,
                         key,
-                        _parse_traces(traces_raw, str(execution_uuid)),
+                        parse_traces(traces_raw, str(execution_uuid)),
                     )
                 )
 
@@ -132,13 +135,13 @@ class SweepStuckExecutionsUseCase:
                 *[str(u) for u, _ in stuck],
             )
             cleanup_pipe.execute()
-        except Exception:
+        except RedisError:
             logger.exception(
                 "[EXEC_LOG] Stuck-sweep cleanup pipeline failed; remnants may linger"
             )
 
         logger.warning(
-            "[EXEC_LOG] Stuck sweep finalised %d execution(s) past the threshold",
-            len(stuck),
+            f"[EXEC_LOG] Stuck sweep finalised {len(stuck)} execution(s) "
+            f"past the threshold"
         )
         return len(stuck)
